Refactor to shared.
